[mod] change project layout, makefile, dockerfile
