+++ example/example_cluster_service_broker.yaml
@@ -2,6 +2,5 @@ apiVersion: servicecatalog.k8s.io/v1beta1
 kind: ClusterServiceBroker
 metadata:
   name: hyperbroker5
-  namespace: cluster-tsb-test
 spec:
-  url: 'http://{tsb_ip}:{tsb_port}'
+  url: 'http://cluster-template-service-broker.cluster-tsb-ns.svc.cluster.local:80'
