+++ example/example_cluster_service_instance.yaml
@@ -1,7 +1,7 @@
 apiVersion: servicecatalog.k8s.io/v1beta1
 kind: ServiceInstance
 metadata:
-  name: nginx
+  name: re-test
   namespace: tsb-test
 spec:
   clusterServiceClassExternalName: cluster-nginx-template
