+++ pkg/server/schemas/common_types.go
@@ -11,5 +11,5 @@ type Error struct {
 	Error            string `json:"error,omitempty"`
 	Description      string `json:"description,omitempty"`
 	InstanceUsable   bool   `json:"instance_usable,omitempty"`
-	UpdateRepeatable bool   `json"update_repeatable,omitempty"`
+	UpdateRepeatable bool   `json:"update_repeatable,omitempty"`
 }
